# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: OrigCMS API
    description: OrigCMS is a modern Content Management System for video platforms.
    contact:
        name: GodCong
        url: https://github.com/origadmin
        email: waitforadding@gmail.com
    license:
        name: MIT
        url: https://origadmin/application/origcms/blob/master/LICENSE
    version: Version from annotation
servers:
    - url: http://localhost:18080
    - url: https://localhost:18080
paths:
...
    /api/v1/me:
        get:
            tags:
                - UserService
            description: GetMe returns the current user's information.
            operationId: UserService_GetMe
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.user.GetMeResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        put:
            tags:
                - UserService
            description: UpdateMe updates the current user's information.
            operationId: UserService_UpdateMe
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.user.UpdateMeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.user.UpdateMeResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
components:
    schemas:
...
        api.v1.services.types.Permission:
            type: object
            properties:
                id:
                    type: string
                create_time:
                    type: string
                    format: date-time
                update_time:
                    type: string
                    format: date-time
                keyword:
                    type: string
                name:
                    type: string
                description:
                    type: string
                type:
                    type: string
                resource_type:
                    type: string
                action:
                    type: string
                sequence:
                    type: integer
                    format: int32
                status:
                    type: integer
                    format: int32
                parent_id:
                    type: string
            description: Permission is the model entity for the Permission schema.
...
        api.v1.services.types.Role:
            type: object
            properties:
                id:
                    type: string
                create_time:
                    type: string
                    format: date-time
                update_time:
                    type: string
                    format: date-time
                keyword:
                    type: string
                name:
                    type: string
                description:
                    type: string
                type:
                    type: integer
                    format: int32
                sequence:
                    type: integer
                    format: int32
                status:
                    type: integer
                    format: int32
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.types.User'
                permissions:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.types.Permission'
            description: Role is the model entity for the Role schema.
...
        api.v1.services.types.User:
            type: object
            properties:
                id:
                    type: string
                create_author:
                    type: string
                update_author:
                    type: string
                create_time:
                    type: string
                    format: date-time
                update_time:
                    type: string
                    format: date-time
                username:
                    type: string
                nickname:
                    type: string
                password:
                    type: string
                phone:
                    type: string
                email:
                    type: string
                avatar:
                    type: string
                status:
                    type: integer
                    format: enum
                is_superuser:
                    type: boolean
                is_staff:
                    type: boolean
                last_login_ip:
                    type: string
                login_ip:
                    type: string
                last_login_time:
                    type: string
                    format: date-time
                login_time:
                    type: string
                    format: date-time
                slug:
                    type: string
                name:
                    type: string
                role:
                    type: string
                is_editor:
                    type: boolean
                is_manager:
                    type: boolean
                title:
                    type: string
                description:
                    type: string
                logo:
                    type: string
                location:
                    type: string
                media_count:
                    type: integer
                    format: int32
                notification_on_comments:
                    type: boolean
                allow_contact:
                    type: boolean
                date_joined:
                    type: string
                    format: date-time
                date_added:
                    type: string
                    format: date-time
                last_login:
                    type: string
                    format: date-time
                profile:
                    $ref: '#/components/schemas/api.v1.services.types.UserProfile'
                setting:
                    $ref: '#/components/schemas/api.v1.services.types.UserSetting'
                roles:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.types.Role'
            description: User is the model entity for the User schema.
        api.v1.services.types.UserProfile:
            type: object
            properties:
                avatar:
                    type: string
                gender:
                    type: string
                name:
                    type: string
                bio:
                    type: string
                website:
                    type: string
                location:
                    type: string
            description: UserProfile holds the profile information of a user.
        api.v1.services.types.UserSetting:
            type: object
            properties:
                theme:
                    type: string
                language:
                    type: string
                timezone:
                    type: string
                preferences:
                    type: object
                    additionalProperties:
                        type: string
            description: UserSetting holds the personalized settings of a user.
...
        api.v1.services.user.GetMeResponse:
            type: object
            properties:
                user:
                    allOf:
                        - $ref: '#/components/schemas/api.v1.services.types.User'
                    description: The current user.
            description: GetMeResponse is the response message for getting current user info.
...
        api.v1.services.user.UpdateMeRequest:
            type: object
            properties:
                nickname:
                    type: string
                    description: The user fields to update.
                avatar:
                    type: string
                bio:
                    type: string
                email:
                    type: string
                phone:
                    type: string
            description: UpdateMeRequest is the request message for updating current user.
        api.v1.services.user.UpdateMeResponse:
            type: object
            properties:
                user:
                    allOf:
                        - $ref: '#/components/schemas/api.v1.services.types.User'
                    description: The updated user.
            description: UpdateMeResponse is the response message for updating current user.
...
        google.protobuf.Any:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        google.rpc.Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/google.protobuf.Any'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
    securitySchemes:
        Basic:
            type: http
            scheme: basic
        Bearer:
            type: apiKey
            name: Authorization
            in: header
...